refactor(settings): extract filename-template shared module and job-polling hook
FilenameTemplateSettings.tsx mixed ~260 lines of types/pure helpers/
constants and two near-identical hand-rolled job polling effects into a
1,020-line component file. Now:

- filenameTemplateShared.ts holds the types, catalogs, and pure helpers
  (component file exports only the component -> react-refresh clean);
  the helper unit tests import from the shared module.
- useSettingsJobPolling(job, buildJobUrl, setJob, onCompleted?) replaces
  both polling effects with fully-declared dependencies (fast 1s cadence
  through the first quarter, then 3s; silent retry on errors preserved).

Component: 1,020 -> 729 lines. Frontend ESLint is now at zero warnings;
typecheck clean; full frontend suite passes (1,461 tests).

Report: reports/codebase-review-2026-07-01.md (M-1 partial, F-2 done)
