+++ frontend/src/components/Settings/FilenameTemplateSettings.tsx
@@ -19,282 +19,40 @@ import {
     Tooltip,
     Typography,
 } from '@mui/material';
-import React, { useEffect, useMemo, useRef, useState } from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { useQuery, useQueryClient } from '@tanstack/react-query';
 import CollapsibleSection from '../CollapsibleSection';
 import { useLanguage } from '../../contexts/LanguageContext';
+import { useSettingsJobPolling } from '../../hooks/useSettingsJobPolling';
 import { Settings } from '../../types';
 import { api } from '../../utils/apiClient';
 import { createTranslateOrFallback } from '../../utils/translateOrFallback';
 import { TranslationKey } from '../../utils/translations';
+import {
+    FilenameTemplateCatalogResponse,
+    FilenameTemplatePreviewResponse,
+    MEDIA_SERVER_EXPORT_OPTIONS,
+    MediaServerExportJob,
+    PRESET_FALLBACK_OPTIONS,
+    PREVIEW_SCENARIO_ORDER,
+    PreviewResult,
+    PreviewScenario,
+    RenameJob,
+    deriveFilenameEffectiveTemplate,
+    deriveFilenamePresetId,
+    getFilenameTemplateWarningMessage,
+    getMediaServerExportErrorMessage,
+    getPresetLabelFallback,
+    mediaServerExportJobUrl,
+    previewResultSignature,
+    renameJobUrl,
+    resolveFilenamePresetSelectValue,
+} from './filenameTemplateShared';
 
 interface FilenameTemplateSettingsProps {
     settings: Settings;
     onChange: (field: keyof Settings, value: any) => void;
 }
-
-type PreviewScenario = 'channel' | 'playlist' | 'single';
-
-const PREVIEW_SCENARIO_ORDER: PreviewScenario[] = ['channel', 'playlist', 'single'];
-
-interface PreviewResult {
-    videoPath?: string;
-    thumbnailPath?: string;
-    subtitlePath?: string;
-    warnings?: Array<{ code: string; message: string }>;
-}
-
-// Two scenarios are equivalent when they render the same paths and warnings.
-// Used to collapse identical preview tabs (e.g. templates that don't reference
-// any source-type-specific token produce the same result for all scenarios).
-function previewResultSignature(result: PreviewResult | undefined): string {
-    return JSON.stringify({
-        v: result?.videoPath || '',
-        t: result?.thumbnailPath || '',
-        s: result?.subtitlePath || '',
-        w: (result?.warnings || []).map((warning) => warning.code),
-    });
-}
-
-export function getFilenameTemplateWarningMessage(
-    warning: { code: string; message: string },
-    t: (key: TranslationKey) => string
-): string {
-    switch (warning.code) {
-        case 'media_playlist_index_unavailable':
-            return t('filenameWarningMediaPlaylistIndexUnavailable');
-        case 'source_collection_metadata_may_be_empty':
-            return t('filenameWarningSourceCollectionMetadataMayBeEmpty');
-        default:
-            return warning.message;
-    }
-}
-
-function getMediaServerExportErrorMessage(
-    error: any,
-    mode: 'off' | 'nfo' | 'nfo_and_source_json',
-    t: (key: TranslationKey) => string
-): string {
-    const code = error?.response?.data?.code;
-    const rawMessage = String(error?.response?.data?.error || '');
-
-    if (code === 'active_downloads') {
-        return t('mediaServerExportActiveDownloadsError');
-    }
-    if (code === 'queued_downloads') {
-        return t('mediaServerExportQueuedDownloadsError');
-    }
-    if (code === 'unsupported_export_mode') {
-        return t(mode === 'off'
-            ? 'mediaServerExportCleanupUnsupportedModeError'
-            : 'mediaServerExportUnsupportedModeError');
-    }
-    if (rawMessage.includes('requires NFO export mode')) {
-        return t(mode === 'off'
-            ? 'mediaServerExportCleanupUnsupportedModeError'
-            : 'mediaServerExportUnsupportedModeError');
-    }
-
-    return t(mode === 'off'
-        ? 'mediaServerExportCleanupError'
-        : 'mediaServerExportRebuildError');
-}
-
-interface FilenameTemplatePresetCatalogItem {
-    id: string;
-    kind: 'legacy' | 'template';
-    labelKey: TranslationKey;
-    descriptionKey: string;
-    template: string;
-    examplePath: string;
-    recommendedSourceTypes: PreviewScenario[];
-}
-
-interface DeprecatedPresetAlias {
-    id: string;
-    labelKey: TranslationKey;
-    mapsToCurrentPresetId?: string;
-}
-
-interface FilenameTemplateReferenceItem {
-    key: string;
-    token: string;
-    descriptionKey: TranslationKey;
-    example?: string;
-    kind: 'liquid' | 'pattern';
-}
-
-interface FilenameTemplateReferenceSection {
-    id: string;
-    titleKey: TranslationKey;
-    descriptionKey?: TranslationKey;
-    items: FilenameTemplateReferenceItem[];
-}
-
-interface FilenameTemplateInformationNote {
-    id: string;
-    textKey: TranslationKey;
-}
-
-interface FilenameTemplateCatalogResponse {
-    presets: FilenameTemplatePresetCatalogItem[];
-    deprecatedPresetAliases: DeprecatedPresetAlias[];
-    informationNotes: FilenameTemplateInformationNote[];
-    referenceSections: FilenameTemplateReferenceSection[];
-}
-
-interface FilenameTemplatePreviewResponse {
-    valid: boolean;
-    errors: string[];
-    resolved: {
-        mode: 'legacy' | 'template';
-        matchedPresetId: string;
-        template: string | null;
-    };
-    previews: Record<PreviewScenario, PreviewResult> | null;
-}
-
-interface RenameJob {
-    id: string;
-    status: 'running' | 'completed' | 'failed' | 'cancelled';
-    lockedAt: number;
-    template: string;
-    total: number;
-    processed: number;
-    succeeded: number;
-    skipped: number;
-    failed: number;
-    currentTitle?: string;
-    items: Array<{
-        videoId: string;
-        title: string;
-        status: string;
-        skipReason?: string;
-        error?: string;
-        oldVideoPath?: string;
-        newVideoPath?: string;
-    }>;
-}
-
-interface MediaServerExportJob {
-    id: string;
-    status: 'running' | 'completed' | 'failed' | 'cancelled';
-    lockedAt: number;
-    mode: 'off' | 'nfo' | 'nfo_and_source_json';
-    action: 'rebuild' | 'cleanup';
-    total: number;
-    processed: number;
-    succeeded: number;
-    skipped: number;
-    failed: number;
-    currentTitle?: string;
-    items: Array<{
-        videoId: string;
-        title: string;
-        status: string;
-        skipReason?: string;
-        error?: string;
-    }>;
-}
-
-const PRESET_FALLBACK_OPTIONS = [
-    { value: 'legacy', labelKey: 'filenamePresetLegacy', fallbackLabel: 'Current compatible mode (Title-Author-Year)' },
-    { value: 'media_center_date_index', labelKey: 'filenamePresetMediaCenterDateIndex', fallbackLabel: 'Media center - Season/Episode by date and index' },
-    { value: 'playlist_static_index', labelKey: 'filenamePresetPlaylistStaticIndex', fallbackLabel: 'Playlist - Season 1 / Episode by index' },
-    { value: 'playlist_static_date', labelKey: 'filenamePresetPlaylistStaticDate', fallbackLabel: 'Playlist - Season 1 / Episode by date' },
-    { value: 'source_date_flat', labelKey: 'filenamePresetSourceDateFlat', fallbackLabel: 'Source - Date then title' },
-    { value: 'custom', labelKey: 'filenamePresetCustom', fallbackLabel: 'Custom template' },
-] as const;
-
-const LEGACY_TEMPLATE = '{{ title }}-{{ uploader }}-{{ upload_year }}.{{ ext }}';
-
-function getPresetLabelFallback(labelKey: TranslationKey): string {
-    switch (labelKey) {
-        case 'filenamePresetLegacy':
-            return 'Current compatible mode (Title-Author-Year)';
-        case 'filenamePresetMediaCenterDateIndex':
-            return 'Media center - Season/Episode by date and index';
-        case 'filenamePresetPlaylistStaticIndex':
-            return 'Playlist - Season 1 / Episode by index';
-        case 'filenamePresetPlaylistStaticDate':
-            return 'Playlist - Season 1 / Episode by date';
-        case 'filenamePresetSourceDateFlat':
-            return 'Source - Date then title';
-        case 'filenamePresetCustom':
-            return 'Custom template';
-        default:
-            return labelKey;
-    }
-}
-
-function resolveVisiblePresetId(
-    presetId: string,
-    deprecatedPresetAliases: DeprecatedPresetAlias[]
-): string {
-    return deprecatedPresetAliases.find((alias) => alias.id === presetId)?.mapsToCurrentPresetId || presetId;
-}
-
-export function deriveFilenamePresetId(
-    settings: Pick<Settings, 'downloadFilenameMode' | 'downloadFilenameTemplate' | 'downloadFilenamePresetId'>,
-    presets: FilenameTemplatePresetCatalogItem[] = [],
-    deprecatedPresetAliases: DeprecatedPresetAlias[] = []
-): string {
-    if (settings.downloadFilenameMode === 'legacy') {
-        return 'legacy';
-    }
-
-    if (settings.downloadFilenameMode === 'template') {
-        const template = settings.downloadFilenameTemplate || '';
-        const matchedBuiltIn = presets.find(
-            (preset) =>
-                preset.kind === 'template' &&
-                preset.template === template
-        );
-        if (matchedBuiltIn) {
-            return matchedBuiltIn.id;
-        }
-        if (settings.downloadFilenamePresetId && settings.downloadFilenamePresetId !== 'custom') {
-            return resolveVisiblePresetId(settings.downloadFilenamePresetId, deprecatedPresetAliases);
-        }
-        return 'custom';
-    }
-
-    return resolveVisiblePresetId(settings.downloadFilenamePresetId || 'legacy', deprecatedPresetAliases);
-}
-
-export function deriveFilenameEffectiveTemplate(
-    settings: Pick<Settings, 'downloadFilenameMode' | 'downloadFilenameTemplate' | 'downloadFilenamePresetId'>,
-    presets: FilenameTemplatePresetCatalogItem[] = []
-): string {
-    if (settings.downloadFilenameMode === 'legacy') {
-        return presets.find((preset) => preset.id === 'legacy')?.template || LEGACY_TEMPLATE;
-    }
-
-    if (settings.downloadFilenameMode === 'template') {
-        return settings.downloadFilenameTemplate || '';
-    }
-
-    const presetId = settings.downloadFilenamePresetId || 'legacy';
-    return presetId === 'custom'
-        ? settings.downloadFilenameTemplate || ''
-        : presets.find((preset) => preset.id === presetId)?.template || '';
-}
-
-export function resolveFilenamePresetSelectValue(
-    derivedPresetId: string,
-    namingMode: 'legacy' | 'template',
-    forceCustomSelection: boolean
-): string {
-    return forceCustomSelection && namingMode === 'template'
-        ? 'custom'
-        : derivedPresetId;
-}
-
-const MEDIA_SERVER_EXPORT_OPTIONS = [
-    { value: 'off', labelKey: 'mediaServerExportModeOff' },
-    { value: 'nfo', labelKey: 'mediaServerExportModeNfo' },
-    { value: 'nfo_and_source_json', labelKey: 'mediaServerExportModeNfoAndSourceJson' },
-] as const;
 
 const FilenameTemplateSettings: React.FC<FilenameTemplateSettingsProps> = ({
     settings,
@@ -344,8 +102,6 @@ const FilenameTemplateSettings: React.FC<FilenameTemplateSettingsProps> = ({
     const [renameError, setRenameError] = useState<string | null>(null);
     const [exportJob, setExportJob] = useState<MediaServerExportJob | null>(null);
     const [exportError, setExportError] = useState<string | null>(null);
-    const pollTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-    const exportPollTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
 
     // Batch rename uses the current form state shown above, not only the last
     // saved defaults used for future downloads.
@@ -408,63 +164,16 @@ const FilenameTemplateSettings: React.FC<FilenameTemplateSettingsProps> = ({
         return () => clearTimeout(timer);
     }, [effectiveTemplate, namingMode, presetId]);
 
-    // Poll rename job
-    useEffect(() => {
-        if (!renameJob || renameJob.status !== 'running') {
-            if (pollTimerRef.current) clearTimeout(pollTimerRef.current);
-            return;
-        }
-        const poll = async () => {
-            try {
-                const res = await api.get<RenameJob>(
-                    `/settings/filename-template/rename-jobs/${renameJob.id}`
-                );
-                setRenameJob(res.data);
-                if (res.data.status === 'completed') {
-                    queryClient.invalidateQueries({ queryKey: ['videos'] });
-                    queryClient.invalidateQueries({ queryKey: ['collections'] });
-                    queryClient.invalidateQueries({ queryKey: ['downloadHistory'] });
-                }
-                if (res.data.status === 'running') {
-                    const delay = res.data.processed < (res.data.total * 0.25) ? 1000 : 3000;
-                    pollTimerRef.current = setTimeout(poll, delay);
-                }
-            } catch {
-                // silently retry
-                pollTimerRef.current = setTimeout(poll, 3000);
-            }
-        };
-        pollTimerRef.current = setTimeout(poll, 1000);
-        return () => {
-            if (pollTimerRef.current) clearTimeout(pollTimerRef.current);
-        };
-    }, [renameJob?.id, renameJob?.status]);
+    // On rename completion the library's paths changed on disk: refresh every
+    // query that renders them.
+    const handleRenameCompleted = useCallback(() => {
+        queryClient.invalidateQueries({ queryKey: ['videos'] });
+        queryClient.invalidateQueries({ queryKey: ['collections'] });
+        queryClient.invalidateQueries({ queryKey: ['downloadHistory'] });
+    }, [queryClient]);
 
-    // Poll media server export rebuild job
-    useEffect(() => {
-        if (!exportJob || exportJob.status !== 'running') {
-            if (exportPollTimerRef.current) clearTimeout(exportPollTimerRef.current);
-            return;
-        }
-        const poll = async () => {
-            try {
-                const res = await api.get<MediaServerExportJob>(
-                    `/settings/media-server-export/jobs/${exportJob.id}`
-                );
-                setExportJob(res.data);
-                if (res.data.status === 'running') {
-                    const delay = res.data.processed < (res.data.total * 0.25) ? 1000 : 3000;
-                    exportPollTimerRef.current = setTimeout(poll, delay);
-                }
-            } catch {
-                exportPollTimerRef.current = setTimeout(poll, 3000);
-            }
-        };
-        exportPollTimerRef.current = setTimeout(poll, 1000);
-        return () => {
-            if (exportPollTimerRef.current) clearTimeout(exportPollTimerRef.current);
-        };
-    }, [exportJob?.id, exportJob?.status]);
+    useSettingsJobPolling(renameJob, renameJobUrl, setRenameJob, handleRenameCompleted);
+    useSettingsJobPolling(exportJob, mediaServerExportJobUrl, setExportJob);
 
     const handlePresetChange = (value: string) => {
         if (value === 'legacy') {
