+++ frontend/src/components/Settings/__tests__/FilenameTemplateSettings.test.tsx
@@ -4,7 +4,7 @@ import {
     deriveFilenamePresetId,
     resolveFilenamePresetSelectValue,
     getFilenameTemplateWarningMessage
-} from '../FilenameTemplateSettings';
+} from '../filenameTemplateShared';
 
 const catalogPresets = [
     {
